pyproject.toml
[project]
name = "nillebcal"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "google-api-python-client>=2.184.0",
    "google-auth>=2.41.1",
    "google-auth-oauthlib>=1.2.2",
    "icalendar>=6.3.1",
    "msal>=1.34.0",
    "pandas>=2.3.3",
    "python-dateutil>=2.9.0.post0",
    "pytz>=2025.2",
    "recurring-ical-events>=3.8.0",
    "requests>=2.32.5",
    "uvicorn>=0.37.0",
]
